Moved the conversation-menu to its own file.
